Changes: 	o edns is added when DO bit is set by pkt2wire. 	o size_t case in rdlen of nsec3 next owner name, for 255 long next 	  owner names. 	o salt string2host from 5+len to 1+len. Fixes, some sort of typo? 	o added NSEC3 keysigning types to the keysize calculation routine. 	o added DO flag (set on returned message) and copy_query to returned 	  message options to ldns-testpackets.

+++ wire2host.c
@@ -221,7 +221,7 @@ ldns_wire2rdf(ldns_rr *rr, const uint8_t *wire, size_t max, size_t *pos)
 		case LDNS_RDF_TYPE_B32_EXT:
 		case LDNS_RDF_TYPE_NSEC3_NEXT_OWNER:
 			/* length is stored in first byte */
-			cur_rdf_length = (uint8_t) wire[*pos] + 1;
+			cur_rdf_length = ((size_t) wire[*pos]) + 1;
 			break;
 		case LDNS_RDF_TYPE_APL:
 		case LDNS_RDF_TYPE_B64:
